requirements.txt
torch>=2.0
torchvision>=0.15
albumentations>=1.3
numpy
Pillow
matplotlib
